feat(pricing): migrate to BigDecimal and implement multi-currency support
- Refactor price fields from double to BigDecimal for financial precision
- Implement currency conversion and display for /api/guests/rooms endpoint
- Fix transaction propagation issue: add explicit @Transactional to override class-level readOnly=true for currency updater
- Update DTOs to use BigDecimal for fullPrice/pricePerNight and add currencyCode field
- Add unit tests for currency conversion logic in GuestService

+++ src/main/java/pl/kacper/reservation/hotelReservationSystem/admin/dtos/RoomRequestDto.java
@@ -5,10 +5,12 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import pl.kacper.reservation.hotelReservationSystem.catalog.RoomEntity;
 
+import java.math.BigDecimal;
+
 public record RoomRequestDto(
         @Min(1) long roomNumber,
         @Min(1) int roomCapacity,
-        @Min(1) double pricePerNight,
+        @Min(1) BigDecimal pricePerNight,
         @NotNull RoomEntity.StandardType standardType
 ) {
 }
