+++ src/main/java/pl/kacper/reservation/hotelReservationSystem/dtos/ReservationResponseDto.java
@@ -2,6 +2,7 @@ package pl.kacper.reservation.hotelReservationSystem.dtos;
 
 import pl.kacper.reservation.hotelReservationSystem.catalog.ReservationEntity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ReservationResponseDto(
@@ -9,6 +10,6 @@ public record ReservationResponseDto(
         ReservationEntity.Status reservationStatus,
         LocalDate checkInDate,
         LocalDate checkOutDate,
-        double fullPrice
+        BigDecimal fullPrice
 ) {
 }
